requirements.txt
Flask
pyyaml
waitress
